+++ src/r_array.rs
@@ -1,6 +1,4 @@
-#[cfg(feature = "friendly-api")]
-use std::iter::FromIterator;
-use std::{cmp::Ordering, convert::TryInto, fmt, os::raw::c_long, ptr::NonNull, slice};
+use std::{cmp::Ordering, fmt, os::raw::c_long, ptr::NonNull, slice};
 
 #[cfg(ruby_gte_3_0)]
 use rb_sys::ruby_rarray_consts::RARRAY_EMBED_LEN_SHIFT;